+++ CHANGELOG.md
@@ -64,6 +64,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 -   Add `GoalExecutionListener` to track goal execution within an SDM.
 -   Add support voting on goal approval in an SDM. [#465](https://github.com/atomist/sdm/issues/465)
 -   Add goal locking model through `LockingGoal` and `Goals.andLock()`
+-   Added `EnforceableProjectInvariantRegistration`
 
 ### Changed
 
@@ -74,6 +75,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 -   **BREAKING**  `CodeTransformRegistration.editMode` is replaced by `transformPresentation`.
 -   **BREAKING** `CommandHandler` registrations must now specify a `listener`. `createCommand` alternative is removed.
 -   **BREAKING** Review listeners must now have names. Introduced `ReviewListenerRegistration`
+-   **BREAKING** `CodeTransformRegistration.react` method renamed to `onTransformResults`
+-   **BREAKING** `CodeInspectionRegistration.react` method renamed to `onInspectionResults`
 
 ### Fixed